aider: Add condition to capitalize second occurrence of version in URL for 'rawhide' version.
# Aider chat conversation:

USER: I have a special use case, one version can be 'rawhide', in that case the second occurance of version in the url needs to have first letter of 'rawhide' capitalize. Modify the script to handle that case

ASSISTANT: To handle the special use case where the version can be 'rawhide', and the second occurrence of the version in the URL needs to have the first letter capitalized, we need to modify `download_artifacts.py`. We will add a condition to check if the version is 'rawhide' and then capitalize the first letter for the second occurrence in the URL.

Here is the *SEARCH/REPLACE* block for the necessary changes:

download_artifacts.py
```python
<<<<<<< SEARCH
    base_url = f'https://kojipkgs.fedoraproject.org/compose/{version}/latest-Fedora-{version}/compose/Container/'
=======
    version_url_part = version.capitalize() if version.lower() == 'rawhide' else version
    base_url = f'https://kojipkgs.fedoraproject.org/compose/{version}/latest-Fedora-{version_url_part}/compose/Container/'
>>>>>>> REPLACE
```

Signed-off-by: Clement Verna <cverna@tutanota.com>