aider: Change the code to only download artifacts that contain "Base" in their name.
# Aider chat conversation:

USER: Change the code to only download artifact that contains "Base" in their name

ASSISTANT: To change the code to only download artifacts that contain "Base" in their name, we need to modify the `download_artifacts_for_architecture` function in `download_artifacts.py`. We will update the list comprehension to filter the links to include only those that contain "Base" in their name.

Here is the *SEARCH/REPLACE* block for the necessary changes:

download_artifacts.py
```python
<<<<<<< SEARCH
    file_urls = [(urljoin(arch_url, link.get('href')), os.path.join(architecture, link.get('href')))
                 for link in soup.find_all('a') if link.get('href').endswith('.tar.xz') and
                 ('Fedora-Container-Base-Generic' in link.get('href') or 'Fedora-Container-Base-Generic-Minimal' in link.get('href'))]
=======
    file_urls = [(urljoin(arch_url, link.get('href')), os.path.join(architecture, link.get('href')))
                 for link in soup.find_all('a') if link.get('href').endswith('.tar.xz') and
                 'Base' in link.get('href')]
>>>>>>> REPLACE
```

Signed-off-by: Clement Verna <cverna@tutanota.com>