aider: Added a new function to decompress .tar.xz artifacts and updated the main function to call it after downloading the file.
# Aider chat conversation:

USER: Now create a template file named Dockerfile.template with the following"

ASSISTANT: To create a new file named `Dockerfile.template`, we will use a *SEARCH/REPLACE block* with an empty `SEARCH` section and the contents of the Dockerfile in the `REPLACE` section. However, you haven't provided the contents that should be included in the `Dockerfile.template`. Please provide the contents you want to have in the `Dockerfile.template`, and I will create the appropriate *SEARCH/REPLACE block* for you.

USER: Add a new function that will decompress the .tar.xz artifacts once downloaded. unpack the xz and then tar file

ASSISTANT: To add a new function that decompresses `.tar.xz` artifacts after they are downloaded, we will modify the `download_artifacts.py` file. The new function will be named `decompress_artifact` and will use the `subprocess` module to call the `tar` command with the appropriate flags to decompress the `.tar.xz` file.

Here is the *SEARCH/REPLACE* block to add the `decompress_artifact` function:

download_artifacts.py
```python
<<<<<<< SEARCH
=======
import tarfile

def decompress_artifact(artifact_path):
    if artifact_path.endswith('.tar.xz'):
        print(f'Decompressing {artifact_path}...')
        # Decompress the .xz file
        subprocess.run(['xz', '-d', artifact_path], check=True)
        # Extract the .tar file
        tar_path = artifact_path.rstrip('.xz')
        with tarfile.open(tar_path) as tar:
            tar.extractall(path=os.path.dirname(tar_path))
        os.remove(tar_path)
        print(f'Decompressed and extracted {artifact_path}')
>>>>>>> REPLACE
```

And we will also need to call this new function after the file is downloaded. Here is the *SEARCH/REPLACE* block to update the `main` function to call `decompress_artifact`:

download_artifacts.py
```python
<<<<<<< SEARCH
                    print(f'Downloaded {data}')
=======
                    print(f'Downloaded {data}')
                    decompress_artifact(filename)
>>>>>>> REPLACE
```

Please note that this implementation assumes that the `xz` command is available on the system where the script is running. If it's not available, you might need to install it or use a different method to decompress `.xz` files.

Signed-off-by: Clement Verna <cverna@tutanota.com>