aider: Fix the issue by passing the correct directory path to `process_artifact` in `decompress_artifact` function.
# Aider chat conversation:

USER: I have got the following error "Traceback (most recent call last):
  File "/var/home/cverna/Documents/LLMs/python-script/download_artifacts.py", line 92, in <module>
    main(version)
  File "/var/home/cverna/Documents/LLMs/python-script/download_artifacts.py", line 58, in main
    decompress_artifact(filename)
  File "/var/home/cverna/Documents/LLMs/python-script/download_artifacts.py", line 84, in decompress_artifact
    process_artifact(artifact_path.rstrip('.tar.xz'))
  File "/var/home/cverna/Documents/LLMs/python-script/download_artifacts.py", line 62, in process_artifact
    with open(index_path, 'r') as index_file:
         ^^^^^^^^^^^^^^^^^^^^^
FileNotFoundError: [Errno 2] No such file or directory: 'aarch64/Fedora-Container-Base-Generic-Minimal.aarch64-40-1.10.oci/index.json'
"

ASSISTANT: The error you're encountering indicates that the script is trying to open an `index.json` file that does not exist at the specified path. This could be due to the `artifact_path` not being correctly modified to point to the directory where the artifact was decompressed.

To resolve this issue, we need to ensure that the `process_artifact` function is called with the correct path to the decompressed artifact directory. We will modify the `decompress_artifact` function in `download_artifacts.py` to pass the correct directory path to `process_artifact`.

Here is the *SEARCH/REPLACE* block to fix the issue:

download_artifacts.py
```python
<<<<<<< SEARCH
        process_artifact(artifact_path.rstrip('.tar.xz'))
=======
        # Ensure we pass the directory path without the file extension
        decompressed_dir = os.path.splitext(artifact_path)[0]
        process_artifact(decompressed_dir)
>>>>>>> REPLACE
```

This change will ensure that the `process_artifact` function receives the correct directory path without relying on string manipulation that might be error-prone. The `os.path.splitext` function is used to safely remove the file extension from the `artifact_path`.

Signed-off-by: Clement Verna <cverna@tutanota.com>