aider: Replace `xz` with `lzma` library for decompression in `download_artifacts.py`.
# Aider chat conversation:

USER: Is there a python library we could use to replace `xz` in the script to decompress the artifact?

ASSISTANT: Yes, there is a Python library called `lzma` which is part of the standard library and can be used to decompress `.xz` files. This library provides support for .xz and legacy .lzma file formats and compression using the LZMA algorithm.

To replace the `xz` command in the `download_artifacts.py` script with the `lzma` library, you would need to modify the `decompress_artifact` function to use `lzma.open` instead of calling the `subprocess.run` with the `xz` command.

Would you like to proceed with this change? If so, please add the `download_artifacts.py` file to the chat so I can provide you with the necessary *SEARCH/REPLACE block* to implement this change.

USER: yes, let's replace xz with the lzma library. Please make sure to update `requirements.txt` if needed to use the lzma library

ASSISTANT: The `lzma` library is part of the Python standard library, so there is no need to update `requirements.txt` for this change. We can proceed to modify the `download_artifacts.py` file to use `lzma` for decompression.

Here is the *SEARCH/REPLACE block* for the change:

download_artifacts.py
```python
<<<<<<< SEARCH
        subprocess.run(["xz", "-d", artifact_path], check=True)
        # Extract the .tar file
        tar_path = artifact_path.rstrip(".xz")
        with tarfile.open(tar_path) as tar:
            tar.extractall(path=os.path.dirname(tar_path))
        os.remove(tar_path)
=======
        # Decompress the .xz file using lzma
        tar_path = artifact_path.rstrip(".xz")
        with lzma.open(artifact_path, 'rb') as compressed, open(tar_path, 'wb') as f:
            shutil.copyfileobj(compressed, f)
        # Extract the .tar file
        with tarfile.open(tar_path) as tar:
            tar.extractall(path=os.path.dirname(tar_path))
        os.remove(artifact_path)
>>>>>>> REPLACE
```

This change will decompress the `.xz` file using the `lzma` library and then extract the `.tar` file using the `tarfile` module, which is also part of the Python standard library. The `shutil.copyfileobj` function is used to write the decompressed data to a `.tar` file. After extraction, the original `.xz` file is removed.

Signed-off-by: Clement Verna <cverna@tutanota.com>