aider: Update Jinja2 `Environment` to use the current path for locating the templates directory.
# Aider chat conversation:

USER: Can you explain me what the following line of code does?     template = env.get_template("Dockerfile")

ASSISTANT: The line of code `template = env.get_template("Dockerfile")` is part of the `process_artifact` function in the `download_artifacts.py` script. This line is responsible for loading a template named "Dockerfile" from the templates directory using the Jinja2 templating engine.

Here's a breakdown of what this line does:

1. `env` is an instance of `Environment` from the Jinja2 library, which has been previously configured to load templates from the "templates" directory.
2. `env.get_template("Dockerfile")` tells Jinja2 to find and load the template file named "Dockerfile" from the configured templates directory.
3. The loaded template is then assigned to the variable `template`.

This template can later be used to render a Dockerfile with dynamic content by substituting placeholders with actual values using the `template.render()` method. The placeholders are defined within the template file using Jinja2's syntax (e.g., `{{ version }}`), and they are replaced with the corresponding values passed to the `render` method.

USER: Can you modify the script so that the `jinja2` `Environment` is using the current path to find the templates directory.

ASSISTANT: To modify the script so that the `jinja2` `Environment` uses the current path to find the templates directory, we need to update the `process_artifact` function in the `download_artifacts.py` file. Specifically, we will change the `FileSystemLoader` to use the current directory's path to locate the `templates` directory.

Here is the *SEARCH/REPLACE block* for the change:

download_artifacts.py
```python
<<<<<<< SEARCH
env = Environment(loader=FileSystemLoader("templates"))
=======
env = Environment(loader=FileSystemLoader(os.path.join(os.path.dirname(__file__), "templates")))
>>>>>>> REPLACE
```

This change will construct the path to the `templates` directory by joining the directory name of the current file (`__file__`) with the "templates" directory name, ensuring that the `Environment` looks for templates in the correct location relative to the script's location.

Signed-off-by: Clement Verna <cverna@tutanota.com>